+++ test/orders.jl
@@ -1,15 +1,16 @@
 @testset "orders" begin
   api_key = "some_api_key"
-  api_secret = "corresponding_api_secret"
+  api_secret = "some_api_secret"
   response = new_order(
     true,
     api_key,
     api_secret,
-    "sell",
+    "buy",
     "btcusd",
     "1",
     "50000.00",
-    "exchange limit"
+    "exchange limit",
+    ["immediate-or-cancel"]
   )
   println(response)
   order_id = parse(Int64, response["order_id"])